Entities adapted due to DB schema validation

+++ criteria/service/src/main/java/eu/cdevreeze/hibernateexperiments/criteria/entity/FilmActorKey.java
@@ -21,5 +21,5 @@ package eu.cdevreeze.hibernateexperiments.criteria.entity;
  *
  * @author Chris de Vreeze
  */
-public record FilmActorKey(Long actorId, Long filmId) {
+public record FilmActorKey(Integer actorId, Integer filmId) {
 }
